src/requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-community
sentence-transformers
faiss-cpu
PyPDF2